+++ src/components/LazyScrollView.tsx
@@ -1,53 +1,12 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import type { LayoutChangeEvent } from 'react-native';
 import Animated, {
-  runOnJS,
-  useAnimatedReaction,
   useAnimatedRef,
   useAnimatedScrollHandler,
   useDerivedValue,
   useSharedValue,
 } from 'react-native-reanimated';
-
-const AnimatedContext = React.createContext({ value: 0 });
-
-export function LazyChild({
-  children,
-  onThresholdPass,
-}: {
-  children: React.ReactNode;
-  onThresholdPass: () => void;
-}) {
-  const threshold = useContext(AnimatedContext);
-  const animatedY = useSharedValue(0);
-  const hasFiredTrigger = useSharedValue(false);
-
-  const handleTrigger = () => {
-    if (!hasFiredTrigger.value) {
-      hasFiredTrigger.value = true;
-      onThresholdPass();
-    }
-  };
-
-  useAnimatedReaction(
-    () => {
-      return threshold.value > animatedY.value;
-    },
-    (hasPassedThreshold) => {
-      if (hasPassedThreshold) {
-        runOnJS(handleTrigger)();
-      }
-    }
-  );
-
-  const onLayout = (e: LayoutChangeEvent) => {
-    if (animatedY.value !== e.nativeEvent.layout.y) {
-      animatedY.value = e.nativeEvent.layout.y;
-    }
-  };
-
-  return <Animated.View onLayout={onLayout}>{children}</Animated.View>;
-}
+import { AnimatedContext } from '../context/AnimatedContext';
 
 export function LazyScrollView({
   children,
